Componentalized a lot of things that will not fit in this commit message
